Make block render when it is used as property
Dummy property is added in block type in order to make the system render
the comments block when it is used as content property. Otherwise block
is considered as empty and it is not rendered in view mode.
Default English names and descriptions are added for block and dynamic
content types for the case when embedded texts are not loaded.